refactor: refactor props to Readonly in multiple components, refactor code to not nest functions more than 4 levels deep
