+++ apps/client/src/features/create-update-question/ui/QuestionModalHeader.tsx
@@ -1,0 +1,9 @@
+function QuestionModalHeader() {
+  return (
+    <header>
+      <h2 className='text-lg font-semibold text-black'>질문하기</h2>
+    </header>
+  );
+}
+
+export default QuestionModalHeader;
