+++ apps/client/src/features/create-update-reply/ui/ReplyModalHeader.tsx
@@ -1,0 +1,9 @@
+function ReplyModalHeader() {
+  return (
+    <header>
+      <h2 className='text-lg font-semibold text-black'>답변하기</h2>
+    </header>
+  );
+}
+
+export default ReplyModalHeader;
