+++ apps/client/src/shared/ui/InputField/InputField.tsx
@@ -17,7 +17,15 @@ const validationStyle: Record<ValidationStatus, string> = {
   INVALID: 'max-h-10 text-rose-500 opacity-100',
 };
 
-function InputField({ label, type, value, onKeyDown, onChange, placeholder, validationStatus }: InputFieldProps) {
+function InputField({
+  label,
+  type,
+  value,
+  onKeyDown,
+  onChange,
+  placeholder,
+  validationStatus,
+}: Readonly<InputFieldProps>) {
   return (
     <div className='flex w-full flex-col items-center'>
       <div className='gap-4r flex w-full flex-row items-center justify-start'>
