+++ apps/client/src/widgets/question-list/ui/QuestionLikeButton.tsx
@@ -2,15 +2,13 @@ import { GrLike, GrLikeFill } from 'react-icons/gr';
 
 import { Button } from '@/shared/ui/button';
 
-function QuestionLikeButton({
-  liked,
-  likesCount,
-  onClick,
-}: {
+interface QuestionLikeButtonProps {
   liked: boolean;
   likesCount: number;
   onClick: () => void;
-}) {
+}
+
+function QuestionLikeButton({ liked, likesCount, onClick }: Readonly<QuestionLikeButtonProps>) {
   return (
     <Button className='hover:bg-gray-200/50 hover:transition-all' onClick={onClick}>
       <div className='flex flex-row items-center gap-2 text-sm font-medium text-gray-500'>
