+++ apps/client/src/widgets/question-list/ui/ReplyCountButton.tsx
@@ -2,7 +2,12 @@ import { RiQuestionAnswerLine } from 'react-icons/ri';
 
 import { Button } from '@/shared/ui/button';
 
-function ReplyCountButton({ count, onClick }: { count: number; onClick: () => void }) {
+interface ReplyCountButtonProps {
+  count: number;
+  onClick: () => void;
+}
+
+function ReplyCountButton({ count, onClick }: Readonly<ReplyCountButtonProps>) {
   return (
     <Button className='hover:bg-gray-200/50 hover:transition-all' onClick={onClick}>
       <div className='flex flex-row items-center gap-2 text-sm font-medium text-gray-500'>
